refactor(query): split server routes and workspace modules
